Abbott Shared HID: document encrypted commands and init sequence.
The Libre 2 is using encrypted commands that are not fully documented yet,
but at least provide information that the length is not valid for them.

(Work on Libre 2 device is part of issue #8.)

Update known details for initialization sequence: `0x01` is the only
command that is actually needed to complete initialization on all my
devices. The other pre-initialization commands appear to exist for the
software to select the correct device.